"1:25pm. I've wasted enough time. Let me resume. It is good to write 'it is time to resume' since it acts as a trigger. If I can figure out homogeneous composition and glue types I've essentially figured out Cubical Agda. I'll be able to study the libraries without much fear at that point.
https://www.reddit.com/r/math/comments/c040x8/homotopy_type_theory_towards_grothendiecks_dream/

https://home.sandiego.edu/~shulman/papers/hott-grothendieck.pdf
https://home.sandiego.edu/~shulman/papers/highlevel.pdf

Let me read these slides by Mike Schulman first. It is kind of difficult to just work up the drive to bash my head against things I do not understand.

1:45pm. `0.5 = 0.499999 · · ·`

I keep seeing this shitty example. Is this actually proven? The whole day I've been thinking how proofs that cannot possibly be constructed - the proofs by 'divergence' are bullshit and now this pops up in an actual paper. Do mathematicians really believe this?

`0.5 and 0.499999 . . . are different decimal expansions, but represent the same real number.`

Clearly something is wonky with real numbers when they lead to such conclusions.

1:55pm.

```
Other “compilers” exist that interpret types as:
* Super-geometric spaces
* Sets with “a compatible action of all groups at once”
* Sets with time-varying dynamical behavior
* Classically observable aspects of quantum systems
* Nonstandard sets containing infinitesimal objects
* Parametrized spectra or excisive homotopy functors
* ∞-categories or (∞, n)-categories
```

The first 3 points (and arguably the 4th) are all machine learning references.

2:20pm. Hmmm, before I resume I suppose you could prove `0.5 = lim (x = 0) 0.5 + x`, by substituting x for 0 and claim that this is the same as `0.5 = 0.49...`, but this is trickery of the worst degree.

Forget it for now.

```
Cubical Agda also has cubical subtypes as in Cohen et al. [2018]; given A : Set ℓ, r : I and
u : Partial r A we can form the type A [ r ↦ u ]. A term v of this type is a term of type A that is
definitionally equal to u when IsOne r is satisfied. Any term u : A can be seen as a term of type
A [ r ↦ u ] that agrees with itself when IsOne r.
```

I keep reading this over and over and it makes zero sense to me. What does a term agreeing with itself mean?

```
_[_↦_] : ∀ {ℓ} (A : Set ℓ) (φ : I) (u : Partial φ A) → Setω
A [ φ ↦ u ] = Sub A φ u
```

Ok, so the `u` inside has to be a `Partial`.

2:30pm. Let me make my question.

```
Cubical Agda also has cubical subtypes as in Cohen et al. [2018]; given A : Set ℓ, r : I and
u : Partial r A we can form the type A [ r ↦ u ]. A term v of this type is a term of type A that is
definitionally equal to u when IsOne r is satisfied. Any term u : A can be seen as a term of type
A [ r ↦ u ] that agrees with itself when IsOne r.
```

I keep reading this over and over and it makes zero sense to me. What does a term agreeing with itself mean?

Does it mean that `v : A [ i1 ↦ u ] = u`? What happens when `r = i0`? This is a pertinent question for me because `outS` will just unwrap it without considering that state of `r`.

`Any term u : A can be seen as a term of type A [ r ↦ u ] that agrees with itself when IsOne r.` This makes no sense because `u` is supposed to be a `Partial`.

This is from the Cubical Agda paper.

```
A term v : A [ φ ↦ u ] should be thought of as a term of type A which is definitionally equal to u : A when IsOne φ is satisfied. Any term u : A can be seen as an term of A [ φ ↦ u ] which agrees with itself on φ:
```

This is the current documentation. A large factor in my confusion over cubical subtypes is that it is saying that `(v : A [ i1 ↦ u ]) = (u : A)`.

2:45pm. Ah, wait, I think I get what it is saying.

It is saying that `A [ i1 ↦ u ] = u`. Really, this should be the first thing that came to my mind.

Does that mean that when `r = i0` that `A [ i0 ↦ u ] = A`.

3pm. No wait, all of this is completely wrong. I think I get it now.

```
inS : {r : I} (a : A) → A [ r ↦ (λ _ → a) ]
outS : {r : I} {u : Partial r A} → A [ r ↦ u ] → A
outS (inS v) = v
inS (outS v) = v
```

Putting things together what `inS` does is map variable `v` from type `A` to `A [ r ↦ (λ _ → v)]`. Meaning `v : A` maps to `v : A [ r ↦ (λ _ → v)]` now. I've left the names the same to highlight that the object is the same and only the type differs. What I wrote does not strictly make sense, but you can imagine variables being tagged at compile time and `u` and `v` (for example) being equal if they have the same tag.

With that in mind, it makes sense that `outS` could just strip out that type information from the type level and get back the same variable. Both `inS` and `outS` are just identity functions under the hood.

What I do not understand yet is why it cannot just be `A [v]`. Why must there be a requirement that `r = i1`?

3:10pm. Let me add a bit to that thread.

3:40pm. I added a bit. Ok, now I at least understand half of what the cubical subtypes are about. The other half should definitely be about proving the definitional equality of different variables.

I need to keep at it."