"9:05am. Let me chill for a while and then I will start.
9:20am. Let me start. I've been slacking for almost 40m already.

9:25am. No really, I need to get going. Let me see what that guy wrote.

```
  hcomp
    {φ = φ ∨ ~ i1 }
    (λ j → λ { (φ = i1) → u i0 1=1
             ; (i1 = i0) → outS u0 })
    (outS u0)
```

I saw this yesterday and thought he might have had edited out by the time today rolled around, but no dice.

Why did he substitute `i` with `i1` in the pattern, but with `i0` in the body.

```
{φ = φ ∨ ~ i1 }
```

I hadn't noticed yesterday that he also substituted it with `i1` in the predicate. Let me start with this. It is really hard to get a grip on what I should be doing when the attempts to explain contain mistakes. This is why it would be good to have many examples and be more explicit. Having many examples allow one to ignore the informal descriptions and focus on what really matters rather than having to read leaves.

9:35am. Actually, now that I've opened the file on my drive I realize he probably just pasted the example where I put in `i0` myself.

10:45am.

---

```
  hcomp
    {φ = φ ∨ ~ i1 }
    (λ j → λ { (φ = i1) → u i0 1=1
             ; (i1 = i0) → outS u0 })
    (outS u0)
```
Did you make a mistake in substituting `i` for `i0` in the first branch `(φ = i1) → u i0 1=1`? I'd expect this to be `(φ = i1) → u j 1=1` once `i1 ∧ j` gets reduced. I am guessing you probably pasted and modified the erroneous example in my last post. I'll go with that assumption for now.

> In particular hcomp u (outS u0) is indeed what hfill u u0 i1 is equal to.

```
hfill {φ = φ} u u0 i =
  hcomp
    {φ = φ ∨ ~ i }
    (λ j → λ { (φ = i1) → u (i ∧ j) 1=1
             ; (i = i0) → outS u0 })
    (outS u0)
```

How could that be the case? Even if you know that `i = i1`, the most you can reduce `hfill` should be...

```
  hcomp
    {φ = φ }
    (λ j → λ { (φ = i1) → u j 1=1
             ; (i1 = i0) → outS u0 })
    (outS u0)
```

This is a far cry from reducing it to `hcomp u (outS u0)`. Even ignoring the implicit argument, I do not see a way of how you could reduce lambda to just `u`.

> Indeed if i = i0 /\ φ = i1 holds then u i0 1=1 = outS u0, that is guaranteed by the type of u0 and by how outS computes, and u i0 1=1 is also the value the whole hcomp ... will reduce to in this case. However, to clarify, this wasn't the case the quoted sentence was talking about.

I think I made a mistake with the order I took in my post. I got sucked into the pace of the paper and said that...

> When `i` is `i0` this is `outS u₀` and when `i` is `i1` it is `hcomp (λ j → λ { (r = i1) → u j 1=1 } u₀`

But really, instead of saying that when i is 'when `i` is `i0` this is `outS u₀`', what I really need to do is consider the `i = i0 /\ φ = i0` and `i = i0 /\ φ = i1` cases separately.

`i = i0 /\ φ = i0`

This just reduces to `outS u0`. And since `φ = i0` the `outS` just strips the `[ φ ↦ u i0 ]` part from `u0`'s type without any additional work being done.

`i = i0 /\ φ = i1`

Here both branches are valid and reduce to `u i0 1=1` and `outS u0` respectively and it is necessary to prove that both of them are definitionaly equal. In the later branch what is now known is that `u0` is of type `A [ i1 ↦ u i0 ]`. So then `outS` can use that as proof that `u0` is definitionally equal to `u i0 1=1`.

Let me also consider the other two possibilities.

`i = i1 /\ φ = i0`

This one is absurd so there is no need to consider it.

`i = i1 /\ φ = i1`

This one reduces to `u j 1=1`.

Actually I still have some questions about this. If I were to interpret as what this is doing, it does seem like `u : ∀ i → Partial i1 A` is acting exactly like a path type here.

...Though in path types, if you have something like `x ≡ y`, you can always refer to `x` and `y` explicitly. Here that seems to have vanished and only the type of the elements is known.

This is a bit confusing as expected. Why isn't there something like a `PartialPath φ x y` in order to have this kind of mechanism resemble regular propositional equality more?

Now that I think about it, I am sure you would need path types to implement a nontrivial function of `∀ i → Partial φ A`.

---

This is the post I've made. There is still a bit left before I can figure this out.

```
hfill' : ∀ {ℓ} {A : Set ℓ} {φ : I}
        (u : ∀ i → Partial φ A) (u0 : A [ φ ↦ u i0 ])
        → outS u0 ≡ hcomp u (outS u0)
hfill' {φ = φ} u u0 i =
  hcomp
    {φ = φ ∨ ~ i }
    (λ j → λ { (φ = i1) → u (i ∧ j) 1=1
             ; (i = i0) → outS u0 })
    (outS u0)
```

Since this thing compiles, I guess he is not lying when he said that 'In particular hcomp u (outS u0) is indeed what hfill u u0 i1 is equal to.'

11am. Ah, now that I look at it more closely, `u` is in fact the same as `λ j → λ { (φ = i1) → u j 1=1 }`. I keep thinking that `hcomp` expects an element of `A` when it actually expects `λ { (φ = i1) → u j 1=1 }` which is a `Partial`.

I see. Well, that explains that. With this I understand `hfill` completely. Let me make one more post and then I will close the issue and have breakfast.

Ugh...sudden thunder out of the blue. I need to run. I'll do what I said later.

11:55am. It came and went. Let me finish the thread. Then come the glue types.

12:15pm. Let me finally commit this. I want to take a proper break here just for a bit.

After that come the glue types."