"5:10pm. I wasn't expecting much just from lying down, but inspiration keeps coming to me. Not just about the next step, but as I prepare to become a real ML researcher, what my principles should be. More than that, what the principles behind the models I will be trying to find should be.
I already had an idea that I should make the reflective property a principle, but in addition to that I should assume that there are various equivalence principles that a learning model should satisfy - such as never forgetting memories. I know that never forgetting seems like insanity on a limited budget, but sometimes making something a core part of your thinking can lead to radical conclusions that do not seem obvious at first glance.

I think I will forget Cubical Agda for the time being. It is a pity that it took me nearly two weeks to just get the basics down - the documentation was really too lacking, but after all this effort, you can bet I will never forget them. I'll be able to hit the ground running once I resume.

One thing I am sure of is that model equivalence are important.

Here is what my current thinking is on some of the properties of a real learning system.

The reflective property is just that a model should be able to approximate its own posterior. I am not sure what the necessary condition for that is, but the sufficient condition is Turing completeness. That means that feedforward nets, conjugate models and tabular agents are right out.

Once you make that a principle some things follow from that.

Namely that you are doing program search. The other thing is that by necessity, whatever the algorithm turns out to be, it will be more complex that the programs it learns. The logic behind that is that `program + learning > program`.

Also, it might be insane to assume this, but learning might be movement between equivalent models.

I always found it interesting what a strong optimization component learning has.

Equivalence sounds like equality, but it is not exact equality. It really does leave a lot of room for optimization. Merge sort is equivalent to insertion sort, but the first is `n * log n` and the other is `n^2` in complexity.

5:30pm. If I really look around me, there are hints everywhere as to what learning should be and should not be.

It seems impossible, but I do not think the chance of me figuring this out is zero if I strain myself enough. It would be impossible if I was on my own, but the force that is driving humanity forward is almost baiting me to come and get it.

Well, fine.

Tomorrow I will begin the first steps towards studying formalization of high school mathematics. I skipped real numbers in Cubical Agda this time around - I'll leave them for later. I can always pick them up if I ever need them.

Tomorrow I will be getting back into Coq and checking out that math library. I have it linked in the last monthly update. Somehow I feel self conscious about that update - I feel like my Reddit handle character is shifting towards edginess. Well it is usually like that, but I am not sure I wanted to be 100% serious about me replacing the entire field of ML. Usually, when I mean ill will I am halfway joking about it. Here I am not."