"1:05pm. Ok, enough wasting my time on 4chan. Let me read that manual.
Focus me. It is just impossible to be motivated to read manuals from the get go. I just need to clear my head and focus on it. I will be done before I know it. 4h is not long and I'll get to playing XCOM after that. Usually I am not so blatant about my desire, but studying libraries which I am going to be doing for the next month or two (or even more) is going to be a snooze fest.

After I am done with these studies, I will be able to finally move to studying ML which should be more interesting.

1:40pm. "make reads the makefile in the current directory and begins by processing the first rule. In the example, this rule is for relinking edit; but before make can fully process this rule, it must process the rules for the files that edit depends on, which in this case are the object files. Each of these files is processed according to its own rule. These rules say to update each ‘.o’ file by compiling its source file. The recompilation must be done if the source file, or any of the header files named as prerequisites, is more recent than the object file, or if the object file does not exist."

I am getting lost in daydreams as I read this, but I am starting to get it.

1:45pm. It seems that it uses dates when deciding whether to update or not. So with one file I am guessing it would always have to update...

No well, with one file `make` would not even be necessary.

I can see it.

```
build: Makefile.coq
    $(MAKE) -f Makefile.coq

Makefile.coq: _CoqProject
    coq_makefile -f _CoqProject -o Makefile.coq

.PHONY: build
```

I understand what this means now. I am not sure if this would run the first action automatically or not, but I know what `PHONY` does.

Let me keep reading.

2:30pm.

"In the previous section we learned that GNU make works in two distinct phases: a read-in phase and a target-update phase (see How make Reads a Makefile). GNU make also has the ability to enable a second expansion of the prerequisites (only) for some or all targets defined in the makefile. In order for this second expansion to occur, the special target .SECONDEXPANSION must be defined before the first prerequisite list that makes use of this feature.

If that special target is defined then in between the two phases mentioned above, right at the end of the read-in phase, all the prerequisites of the targets defined after the special target are expanded a second time. In most circumstances this secondary expansion will have no effect, since all variable and function references will have been expanded during the initial parsing of the makefiles. In order to take advantage of the secondary expansion phase of the parser, then, it’s necessary to escape the variable or function reference in the makefile. In this case the first expansion merely un-escapes the reference but doesn’t expand it, and expansion is left to the secondary expansion phase. For example, consider this makefile:"

Well the manual did warn me to only read the first few chapters of every section. I can't deal with this shit right now.

2:35pm. Forget this. I need a break.

2:55pm. Let me resume.

3:30p. I am finally on chapter 6. I am not focusing deeply on this at all. Really, I knew it would go this way. I should have just stopped at the first chapter rather than force myself to do this, but I feel that it is my responsibility to at least skim the manual once. Since I am going to be using `make`, it needs to be this way.

Forget about reading the descriptions. Let me just focus on the examples. That is the way programming is learned.

Ideally you would go from descriptions to examples, but humans brains were not wired like that. I won't have the motivation to tackle the descriptions without getting confused by the examples first.

It is always like that.

4:25pm. I am an idiot for doing this. I spent maybe 20% of the time actually thinking about the `make` - it seems I got quite a lot from the example and the first section of the manual, but the rest after that was filler.

As for the other 80%, a quarter of it I spent surfing 4chan in distraction, but the other 75% I thought about my dreams.

This is actually the usual process for me. It is easy to tell yourself to do this or that, and then not do it. Typically where people fail is the motivation building aspect of doing work so they have to be cajoled into everything.

Motivation to do a task is like a muscle. It can be exercised. In fact it must be exercised. Before a task can be attempted, the motivations need to be elaborated.

I guess the manual is like an excuse for me to do that. Sometimes watching lectures takes the place of that too. I am not quite ready to do what I want so before it can even be attempted, the dreaming is necessary.

Why is Spiral not enough? Why do I need advanced type theory? Just what do I expect the real thing to be? What principles should I adopt? What are the pieces that need to be tied together?

Necromutation is not good enough. It is good enough for the ones in the story, but they have mind uploading, the brain cores and the elaborated theory of learning.

The task of figuring out the theory of learning is enormous. If I can do it, I can consider myself a genius.

That word used to mean divine inspiration, and I am certainly going to need some of that in order to do it. I think just the fact that I have enough courage to attempt this is proof that my subconscious has some inkling of the truth. It definitely is not all there yet though.

It is still asking for more data.

4:40pm. It is nearly 5pm now. Let me try out the examples in the SO post. I've wasted enough time on the manual. After that I'll call it a day.

`make` is simple enough, but I am reminded of the old saying that either a build tool never becomes popular enough to be used, or it becomes a fully fledged PL. It is stupidly complex beyond the basics as evidenced by the sheer size of the manual.

5:20pm. No. This is not working for me. It does seem like `coq_makefile` is generating the file...

5:25pm. Ugh, I accidentally ran `coq_makefile` in the `vfa` folder and accidentally overwrote the ones I had before.

```
coq_makefile -f _CoqProject -o Makefile
```

Rather than doing what the SO answer says, I should run `coq_makefile` directly on the project file. But this is not enough. Now that I am looking at the file, there are some notable differences from the `vfa` one. There has to be some kind of option that I am missing. I should look at the `coq_makefile` documentation carefully.

5:45pm. "From my experimentation, I am not convinced that `-include Makefile.coq` is enough to enable PG functionality. Or that the VFA `Makefile` was generated by `coq_makefile -f _CoqProject -o Makefile`. If you look at the VFA `Makefile`, you will see that that it explicitly has the included files like `Perm.v`. Comparing it with the generated file from the directory's `_CoqProject`, it is obvious that they are different. Are there some compilation options we are missing here? I suppose they could have always put in the files by hand, but I find that unlikely."

Good thing I did not accept the answer too early. I can't believe that I am still dealing with this. Maybe I should try mailing Appel directly?

5:55pm. Ah, I always regret it when I spend a day essentially doing nothing. Tomorrow, I should at least be able to compile Coquelicot library. After that I'll start studying it.

I'll start from the bottom and work my way up. It is time I start warming up to this great task. I can't waste my time going around it anymore. The makefile thing will get resolved eventually.

6pm. Now, XCOM."