"3:35pm. Oh yeah, this reminds me, maybe I should consider finding the book by Adam Chipatala on Coq.
I heard it is for advanced users, so it might cover the advanced rewriting as seen in the library.

4:35pm. There was so more back and forth in the issue. I finally got everything resolved.

4:50pm. No, actually I was too in a hurry. Even though I can get both VFA and Coquelicot to compile just fine, I can't get Emacs + PG, to compile the file properly.

5:10pm. Reopened the issue. It seems VFA works only because all the files are in the same directory.

5:15pm. You can really see how much time these build issues can take up. I hadn't even looked at the manual in the past 1.5 hours.

Let me read it for a bit and I will call it a day.

5:55pm. https://coq.inria.fr/refman/proof-engine/tactics.html

I'll go through this tomorrow. I got into it for a bit just now, but now it is time to call it a day. I am having so much trouble just starting the task. But either way, the build issues are almost done with so tomorrow I should be able to focus on studying the actual proofs and the documentation.

6pm. Lunch, then XCOM in that order. I've earned at least that much."