"10:40am. I went to bed at 3:40am yesterday. I think this is probably not only the most late gaming session is in the last 4.5 years, but probably in my life. I just shows how starved I am for gaming time. I stop the day period at 6pm, then after lunch and bath, it was 7:30pm. If I go to bed at 9pm as is usual for me that means I only have 1.5 hours of game time. Even going to bed at 11pm would not sate that need.
That having said, I was absolutely surprised at the time when I looked at it. The session was just so good. I was using fatigue as a measure of when to stop, but even so the only reason I stopped when I did is because the game crashed and ejected me to desktop.

New rule: start looking at the time. Had I done so, no way would I have gone past 1am. In fact I'd probably have stopped at 12am.

10:50am. Well then. I guess it is time I get started on studying the Coq documentation. I am fatigued, and I'd rather play games than study, but the satisfying session gives me energy to march another day.

10:55am. Everything will one day end in a bad way if I do not succeed at this. Though it is unfair, the games won't give me the power to meet life's challenges on their own.

I have to master the Coq tactics as an intermediate step to even studying the library. Bit by bit I will do it."