"11:10am. "If a minus sign is given between at and the list of occurrences, it negates the condition so that the clause denotes all the occurrences except the ones explicitly mentioned after the minus sign."
It is difficult to understand all of this just from reading the manual. PL work best with plenty of examples.

So far I haven't gotten a reply, but pivoting like I did yesterday is a bad thing to do. I accepted the guy's answer and gave him his reward, but then I turned around and plopped work into his lap again. Maybe it would be better to open a separate question a few days from now. That is what a sensible person would have done.

Though it is always possible that I'll get an answer anyway. At any rate, it does not matter too much. I can always just do the above later.

11:15am. "Here are some tactics that understand occurrence clauses: set, remember, induction, destruct."

I really am going to have to ask for examples if the docs do not offer them.

11:35am. Oh, this damn chapter. I thought it might have been just me, but the way information is presented here I cannot grok at all. Let me move to the next one. Maybe it is best that I skim all of this and then move to the book by Adam, and complement the rest with SO questions.

11:45am. https://coq.inria.fr/refman/proof-engine/ltac.html#goal-selectors

I can understand this thing in particular.

11:55am. https://coq.inria.fr/refman/proof-engine/detailed-tactic-examples.html

Let me leave this for a bit later. I'll take a break here. It won't take me too long to go through the documentation at this rate.

12:30pm. Let me resume. I do not want to waste too much time today during the break. Let me go through the manual.

12:40pm. This stupid thing is just on how induction works. It is incredibly opaque and I do not feel like reading it too closely.

12:45pm. https://coq.inria.fr/refman/proof-engine/ssreflect-proof-language.html

I am finally here. But I do want to take a bigger break regardless.

1:05pm. https://coq.inria.fr/refman/proof-engine/ssreflect-proof-language.html

The intro to this is interesting. I had no idea the tactics I was looking at are actually partly a different language.

Let me stop this for a bit. I want to watch a talk by Gonthier.

https://www.youtube.com/watch?v=_NDD_jXGwk8
Georges Gonthier: The Logic of Real Proofs

He and his team came up with SSReflect and I saw his name come up before, so I might as well get a tad familiar with him.

2pm. https://youtu.be/_NDD_jXGwk8?t=2036
"...So it's a moderately large proof. And its still rather manageable. It is two real screenfuls. You can see the entire proof if you have a reasonable desktop setup. If you go about it the wrong way, then the proof becomes significantly larger. And in fact, I have a case in point because as it happens this theorem was needed in the proof of the Kepler conjecture. So it was reformalized in HOL Light. And this is just part of the proof in HOL Light. Even if in principle the proof language in HOL Light is about as compact as the proof languages I use in Coq, in practice that proof is about 20 times larger.

There are cases where it appears you just have the wrong formalism."

2:50pm. Done with the talk. It was boring, but illustrative of just how far I need to go for real math mastery.

Also I finally closed the issue for good. It seems he does not know the answer so he suggested we move the discussion to PG issues or the discourse forum. It is fine. I had enough of this for the time being. I'll do it later when I need it.

Now let me take a slight break and I will go for the documentation.

3pm. Yeah, I need a break. Out of all the parts of programming, reading documentation is my least favorite part. I really hope that SSReflect turns out to be not too complex.

It is really impossible to be enthusiastic about reading manuals, especially the low example ones like Coq.

3:25pm. Let me resume. I will finish the SSreflect chapter. Let me open a script.

```
occ_switch ::= { [+ | -] num* }
```

I wonder if the `+` and `-` here mean the things to the right or the left of the number? I am thinking that this might have something to do with th rewrite.

3:50pm. https://www.reddit.com/r/haskell/comments/cr80iz/dependent_types_a_pessimist_debates_an_optimist/
https://news.ycombinator.com/item?id=20714137

Let me read this thread while I have some watermelon.

4:20pm. Let me get back into the action. I'll finish the damn chapter today at least. Since I did not get good enough sleep I am having trouble focusing (even moreso than I would have usually for this kind of material), but nonetheless I am going to put in another hour into this.

4:35pm. "While novice or casual Coq users may find the automatic name selection feature convenient, the usage of such a feature severely undermines the readability and maintainability of proof scripts, much like automatic variable declaration in programming languages. The SSReflect tactics are therefore designed to support precise bookkeeping and to eliminate name generation heuristics."

I did not expect to see them signal this. I've made up my mind. SSReflect will be worth adopting if nothing else, just for this.

5pm.

```
Lemma subnK : forall m n, n <= m -> m - n + n = m.
Proof. move=> m n le_n_m.
       elim: n m le_n_m => [| n IHn] m => [_ | lt_n_m].
```

I feel a little spark when I look at that last line, but I am too fatigued to go into the proper deep focus that I need to internalize this. After grinding so much through tedious nonsense, I finally have something that seems like it will be worth learning.

5:05pm. I can't follow this. I find that the information is just too sparse, though my lack of focus might be playing a part in this. I'll be really surprised if today's gaming session goes on anywhere nearly as long as it had yesterday.

Yesterday was great - I was so rested that day that it is a pity I spent all of the working day messing with build issues. Right now is when I needed that focus. Eah. I am just so unlucky that things would go like this.

5:10pm. https://coq.inria.fr/refman/proof-engine/ssreflect-proof-language.html#the-apply-tactic

I am thinking of stopping here. I want to say that I will resume from this point, but I am going to have to go through the chapter from the start tomorrow.

Let me skim it whole and then I'll look for the Coq book by Adam.

http://adam.chlipala.net/cpdt/html/toc.html
Certified Programming with Dependent Types

Ah, it is free and online.

5:30pm. Let me call it a day here. I said I would skim it and I did, but while I was doing that it occurred to me just how long this chapter is. 169 pages according to print preview.

This is going to take a while. It seems now that two books dropped into my lap today. I'll be lucky if I managed to finish this by the end of the month. Even just the task of starting the study of the library seems like it will take a while. Things always turn out like this. On the bright side what I've read so far of SSReflect does make me motivated to study it. I never felt at home with Coq's design and replacing my current set of tactics with different ones would go a long way with improving my skill at it.

As ugly as tactics are, I do feel like they have the usability to make up for it."