"10:05am. It is starting to get dull. Right now my guys are all leveled up and I am just walking all over the aliens. I think I will just rush the story missions from here on out. I do not even need the rest of the tech tree. Damn it, this always happens. It is a few days of exquisite bliss and then the grind starts to set in.
10:15am. And speaking of grind, I think it is time I resume studies on the library. I am really damn groggy right now though, so it will have to be in a while.

10:50am. I am over my groginess a little now. Let me finally start. What I will do to start things off is ask about the reals in Cubical Agda.

https://stackoverflow.com/questions/28404520/how-to-define-real-number-in-agda

11:05am. https://stackoverflow.com/questions/57636512/how-would-reals-be-implemented-in-cubical-agda

I made this last one.

https://hal.inria.fr/hal-00806920v1/document
Formalization of Real Analysis: A Survey of Proof Assistants and Libraries

Before I start studying the library directly let me read this.

11:15am. "The Coq library is structured into two parts: the initial library, which contains elementary logical notions and data-types, and the standard library, a general-purpose library containing various developments and axiomatizations about sets, lists, sorting, arithmetic, real numbers, etc. The standard library reals are axiomatic and described in Section 3.1.2.

Recent developments in Nijmegen led to a different library called C-CoRN (CruzFilipe et al. 2004) and its successor MathClasses (Krebbers and Spitters 2013). The main idea is to provide Bishop-like constructive mathematics. These libraries are described in Section 3.2.3."

I'll have to remember to look into this. Since Coquelicot uses the standard library definitions I'll forget it for now.

"The PVS system and the standard library of Coq both postulate real numbers as a
complete Archimedean field."

What is a 'complete Archimedean field'.

12:10pm. Finished reading that paper. As expected of things that cannot really be computed, they are retardedly complex.

Mathematics is a shitshow with veneer of seriousness on top. All the good stuff seems to have been taken by programming. Yeah, I know I suck at it, but it is not like those that don't can create AI which is the only game worth playing.

I really feel that I got a lot from Spiral, but now I have to throw that away in order to embark on a risky venture. I want to say that it is pointless, but it is the exact opposite of that. It is the only thing that has a point out of everything that I could do.

12:15pm. From the paper, it seems the Coq standard library definitions are the only ones that directly use axioms to represent reals.

This is actually how reals are used in the real world. You make up some rules that you expect the constructions would obey if they were finite and pretend they are valid. If it turns out that the rules are consistent, then the reasoning works out. It is all ad-hoc, but it also works. And it is much easier than the pointless formal reasoning.

It does not really matter what the constructions are if they cannot be computed, so axioms might as well be used.

12:20pm. Nonetheless, I am going to have to study the Coq standard library before I can move to Coquelicot.

Let me take a break here. I'll dive into the library proper after that.

I think it won't be too difficult from here on out. I broke the ice with Cubical Agda. And I am breaking it with SSReflect. In a month I should be fairly familiar with it and should be able to drastically pick up the pace.

Once I have a grasp of how high school math can be verified on a computer that will open the door to the next steps for me.

12:25pm. I think that in order to accomplish what I want to do, I am going to have to learn some pretty advanced mathematics either way. The only way that will ever be possible if I can get the cubical type theory to give me a hand.

At the end of all this, I will fuse all the pieces together and create an actual learning system that makes sense in its workings.

Right now it feels like math has been fractured into dozens of separate fields. This is a consequence of inadequate foundations.

12:30pm. I really with I was born 30 years from now and some guy in my place has already worked all of this out. I am the kind of person who would rather play games rather than make them. Things never go the way I want them to. I never get the battle that I seek. Yet, the battle I have now is the one that will have to be won.

I'll do it. Once I get through the dumb intro stages that I've been going through, things will get more fun."