"https://news.ycombinator.com/item?id=20714137
I really am enjoying myself reading this thread a bit too much. Maybe I will consider pron's series on TLA at some point.

Let me watch Granbelm and I will resume.

2:05pm. I got a reply by Kovacz saying that the reals I am asking about require indexed HITs which Cubical Agda does not have at the moment.

https://arxiv.org/pdf/1902.00297.pdf
Signatures And Induction Principles For Higher Inductive-Inductive Types

In response, I asked him a really difficult question which is what indexed HITs are. I also asked that of Google and it pointed me at this paper for whom he is one of the authors.

2:10pm. No, this paper is too much for me. Instead let me read the section he referred me to on page 393 of the HoTT book and I will start studying the Coq library for the reals.

I need to go through this. This is my gauntlet.

2:35pm. I had my fill of the book. Obviously I do not understand most of it. How much I understand the HoTT book is probably a decent approximation of how well I understand math in general. If I ever get to the level where I can honestly state that I understand pretty much everything in it without too much effort, then I will know I have made it as a mathematician.

Let me take a little break and then I will get started on studying the library.

2:50pm. https://pron.github.io/posts/people-dont-write-programs

Let me read this and then I will start.

3pm. https://pron.github.io/posts/correctness-and-complexity

Let me read this as well.

"This mean that there are small programs — just a couple hundred lines or less — whose behavior cannot be mathematically determined. Not because the analysis would be too large or intractable, but because mathematics is not powerful enough to ever prove the behavior of this program, even in theory."

"Programs are fundamentally and essentially at odds with full understanding, proof and correctness.

As mathematics — as an academic discipline, not as a tool — is concerned with what’s inside the small circle, and computer science is largely concerned with what’s outside it — and, as we’ll see, even those programs that do lie within that inner circle, many are just provable in principle, but not feasibly provable — I think it is clear that computer science isn’t math"

This guy is pretty interesting.

Agh, yeah I get it. This is why I did not want to go into math in the first place. I have no choice, but to go down this route. I know a full fusion is something fairly extreme.

3:20pm. "By the way, the state space of the towers of Hanoi program is a Sierpinski triangle!"

3:25pm. Done. Let me finally start this thing."